+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 - attachments.php created and added to functions.php
 - fields corresponding to slider created and required arguments passed in attachment.php
 - retrieve slider to the theme
-
+- check if attachment class exists else show thumbnail
+- show testimonial with attachment plugin
+- show team members with attachment plugin